1_Automated_Catalog_QC/requirements.txt
# Core ML and AI libraries
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.20.0
Pillow>=9.0.0
numpy>=1.21.0

# Web framework
Flask>=2.2.0
Flask-CORS>=3.0.10

# HTTP requests
requests>=2.28.0

# Utilities
python-dotenv>=0.19.0
